Focused on Localization for Taiwanese People ...
Focused on Localization for Taiwanese People and fixed many mistranslations.
Basic on the original TC translations, this change can help Taiwanese people better know what the strings mean.
What have I changed? e.g. Still in Simplified: 后→後, 鬆鼠→松鼠, 虫→蟲 ... and many other., Localization not for Taiwan(zhTW): 兼容→相容, 性能→效能, 激活→啟動, and so on. Integration for Naga the boss: 娜迦、納迦→那伽 (According to WikiPedia.)
Sum it up, I only changed it for localization, not for new-added strings.