Move hard-coded range values for Magic Trees into configurations.
I've arbitrarily set minimums and maximums, although I don't believe
these are too extreme. I've tested the sorting tree the most, as it's
the easiest to test. All of the others appear to continue to function
with their default values.

Perhaps this has consequences that I'm unable to see, but at the very
least having the ability to enable/disable certain effects in the
config is highly desirable.

Apologies if the indentation is wrong, my editor is playing up.