Tentatively convert Sortingwood to use CAPABILITY_ITEM_HANDLER.
I've no idea if I've done it properly or not, but it seems to function
correctly with chests (& double chests, I believe that's compensated for
by only having the single handler shared between the two chests) and
Iron Chests.

As I've specifically fetched the handler for the "top" slot, it doesns't
take anything out of nearby furnaces, but there's every chance that the
code could mess around with the input slots of normal machines.

If it does, I'm not sure how to fix this. :-?