Additional loot tables as per #741
I'm not sure that I haven't messed up. For the Tower Ghast, I'm
uncertain if it would by default automatically go to the Minecraft Ghast
loot table -- I didn't see anything that would override it.

Otherwise, where a loot table was not already previously specified or
assumed due to inheritance, a blank one was created.

Where there was an existing vanilla one, I copied the example of bunny.
Where there was an inheritance from a Twilight Forest entity, I copied
referenced that directly.