Uncrafting shift-click fix
While in the uncrafting menu, shift-clicking on an item in your inventory now transfers it into the uncrafting input slot. Shift clicking on an item in the uncrafting result matrix now properly only gives the player the item available without depleting the rest of the items in the uncrafting input slot.

Additionally, custom uncrafting recipe logic is slightly changed to probably be better.

I can't find any more bugs.