Update to the French localisation files
The French localization files of the Twilight Forest have not been updated in what I presume to be years. This proposition here is the first part of an uptate/fix/general improvement to the the French localization files.

While still not complete, I do intent to continue onward and finish this if my pull request is accepted into the near future.